+++ vscode/nullc_lang_server/document.h
@@ -4,5 +4,6 @@
 
 struct Document
 {
+	std::string uri;
 	std::string code;
 };