+++ vscode/nullc_lang_client/CHANGELOG.md
@@ -3,3 +3,7 @@ All notable changes to the "nullc" extension will be documented in this file.
 
 ## [Unreleased]
 - First experiments with an extension: default language configuration, simple text-mate syntax and native language server connection
+- Added fold regions
+- Added hovers
+- Added document symbols (namespaces and functions)
+- Added document diagnostics